+++ yash-env/src/system/real/signal.rs
@@ -30,7 +30,7 @@ use std::ops::RangeInclusive;
 #[must_use]
 fn rt_range() -> RangeInclusive<RawNumber> {
     #[cfg(target_os = "aix")]
-    return nix::libc::SIGRTMIN..=nix::libc::SIGRTMAX;
+    return libc::SIGRTMIN..=libc::SIGRTMAX;
 
     #[cfg(any(
         target_os = "android",
@@ -38,7 +38,7 @@ fn rt_range() -> RangeInclusive<RawNumber> {
         target_os = "l4re",
         target_os = "linux",
     ))]
-    return nix::libc::SIGRTMIN()..=nix::libc::SIGRTMAX();
+    return libc::SIGRTMIN()..=libc::SIGRTMAX();
 
     #[allow(unreachable_code)]
     {
@@ -56,17 +56,17 @@ impl Name {
         }
 
         match self {
-            Self::Abrt => wrap(nix::libc::SIGABRT),
-            Self::Alrm => wrap(nix::libc::SIGALRM),
-            Self::Bus => wrap(nix::libc::SIGBUS),
-            Self::Chld => wrap(nix::libc::SIGCHLD),
+            Self::Abrt => wrap(libc::SIGABRT),
+            Self::Alrm => wrap(libc::SIGALRM),
+            Self::Bus => wrap(libc::SIGBUS),
+            Self::Chld => wrap(libc::SIGCHLD),
             #[cfg(any(
                 target_os = "aix",
                 target_os = "horizon",
                 target_os = "illumos",
                 target_os = "solaris",
             ))]
-            Self::Cld => wrap(nix::libc::SIGCLD),
+            Self::Cld => wrap(libc::SIGCLD),
             #[cfg(not(any(
                 target_os = "aix",
                 target_os = "horizon",
@@ -74,7 +74,7 @@ impl Name {
                 target_os = "solaris",
             )))]
             Self::Cld => None,
-            Self::Cont => wrap(nix::libc::SIGCONT),
+            Self::Cont => wrap(libc::SIGCONT),
             #[cfg(not(any(
                 target_os = "android",
                 target_os = "emscripten",
@@ -83,7 +83,7 @@ impl Name {
                 target_os = "linux",
                 target_os = "redox",
             )))]
-            Self::Emt => wrap(nix::libc::SIGEMT),
+            Self::Emt => wrap(libc::SIGEMT),
             #[cfg(any(
                 target_os = "android",
                 target_os = "emscripten",
@@ -93,9 +93,9 @@ impl Name {
                 target_os = "redox",
             ))]
             Self::Emt => None,
-            Self::Fpe => wrap(nix::libc::SIGFPE),
-            Self::Hup => wrap(nix::libc::SIGHUP),
-            Self::Ill => wrap(nix::libc::SIGILL),
+            Self::Fpe => wrap(libc::SIGFPE),
+            Self::Hup => wrap(libc::SIGHUP),
+            Self::Ill => wrap(libc::SIGILL),
             #[cfg(not(any(
                 target_os = "aix",
                 target_os = "android",
@@ -105,7 +105,7 @@ impl Name {
                 target_os = "linux",
                 target_os = "redox",
             )))]
-            Self::Info => wrap(nix::libc::SIGINFO),
+            Self::Info => wrap(libc::SIGINFO),
             #[cfg(any(
                 target_os = "aix",
                 target_os = "android",
@@ -116,7 +116,7 @@ impl Name {
                 target_os = "redox",
             ))]
             Self::Info => None,
-            Self::Int => wrap(nix::libc::SIGINT),
+            Self::Int => wrap(libc::SIGINT),
             #[cfg(any(
                 target_os = "aix",
                 target_os = "android",
@@ -128,7 +128,7 @@ impl Name {
                 target_os = "nto",
                 target_os = "solaris",
             ))]
-            Self::Io => wrap(nix::libc::SIGIO),
+            Self::Io => wrap(libc::SIGIO),
             #[cfg(not(any(
                 target_os = "aix",
                 target_os = "android",
@@ -141,13 +141,13 @@ impl Name {
                 target_os = "solaris",
             )))]
             Self::Io => None,
-            Self::Iot => wrap(nix::libc::SIGIOT),
-            Self::Kill => wrap(nix::libc::SIGKILL),
+            Self::Iot => wrap(libc::SIGIOT),
+            Self::Kill => wrap(libc::SIGKILL),
             #[cfg(target_os = "horizon")]
-            Self::Lost => wrap(nix::libc::SIGLOST),
+            Self::Lost => wrap(libc::SIGLOST),
             #[cfg(not(target_os = "horizon"))]
             Self::Lost => None,
-            Self::Pipe => wrap(nix::libc::SIGPIPE),
+            Self::Pipe => wrap(libc::SIGPIPE),
             #[cfg(any(
                 target_os = "aix",
                 target_os = "android",
@@ -160,7 +160,7 @@ impl Name {
                 target_os = "nto",
                 target_os = "solaris",
             ))]
-            Self::Poll => wrap(nix::libc::SIGPOLL),
+            Self::Poll => wrap(libc::SIGPOLL),
             #[cfg(not(any(
                 target_os = "aix",
                 target_os = "android",
@@ -174,7 +174,7 @@ impl Name {
                 target_os = "solaris",
             )))]
             Self::Poll => None,
-            Self::Prof => wrap(nix::libc::SIGPROF),
+            Self::Prof => wrap(libc::SIGPROF),
             #[cfg(any(
                 target_os = "aix",
                 target_os = "android",
@@ -186,7 +186,7 @@ impl Name {
                 target_os = "redox",
                 target_os = "solaris",
             ))]
-            Self::Pwr => wrap(nix::libc::SIGPWR),
+            Self::Pwr => wrap(libc::SIGPWR),
             #[cfg(not(any(
                 target_os = "aix",
                 target_os = "android",
@@ -199,8 +199,8 @@ impl Name {
                 target_os = "solaris",
             )))]
             Self::Pwr => None,
-            Self::Quit => wrap(nix::libc::SIGQUIT),
-            Self::Segv => wrap(nix::libc::SIGSEGV),
+            Self::Quit => wrap(libc::SIGQUIT),
+            Self::Segv => wrap(libc::SIGSEGV),
             #[cfg(all(
                 any(
                     target_os = "android",
@@ -210,7 +210,7 @@ impl Name {
                 ),
                 not(any(target_arch = "mips", target_arch = "mips64", target_arch = "sparc64"))
             ))]
-            Self::Stkflt => wrap(nix::libc::SIGSTKFLT),
+            Self::Stkflt => wrap(libc::SIGSTKFLT),
             #[cfg(not(all(
                 any(
                     target_os = "android",
@@ -221,24 +221,24 @@ impl Name {
                 not(any(target_arch = "mips", target_arch = "mips64", target_arch = "sparc64"))
             )))]
             Self::Stkflt => None,
-            Self::Stop => wrap(nix::libc::SIGSTOP),
-            Self::Sys => wrap(nix::libc::SIGSYS),
-            Self::Term => wrap(nix::libc::SIGTERM),
+            Self::Stop => wrap(libc::SIGSTOP),
+            Self::Sys => wrap(libc::SIGSYS),
+            Self::Term => wrap(libc::SIGTERM),
             #[cfg(target_os = "freebsd")]
-            Self::Thr => wrap(nix::libc::SIGTHR),
+            Self::Thr => wrap(libc::SIGTHR),
             #[cfg(not(target_os = "freebsd"))]
             Self::Thr => None,
-            Self::Trap => wrap(nix::libc::SIGTRAP),
-            Self::Tstp => wrap(nix::libc::SIGTSTP),
-            Self::Ttin => wrap(nix::libc::SIGTTIN),
-            Self::Ttou => wrap(nix::libc::SIGTTOU),
-            Self::Urg => wrap(nix::libc::SIGURG),
-            Self::Usr1 => wrap(nix::libc::SIGUSR1),
-            Self::Usr2 => wrap(nix::libc::SIGUSR2),
-            Self::Vtalrm => wrap(nix::libc::SIGVTALRM),
-            Self::Winch => wrap(nix::libc::SIGWINCH),
-            Self::Xcpu => wrap(nix::libc::SIGXCPU),
-            Self::Xfsz => wrap(nix::libc::SIGXFSZ),
+            Self::Trap => wrap(libc::SIGTRAP),
+            Self::Tstp => wrap(libc::SIGTSTP),
+            Self::Ttin => wrap(libc::SIGTTIN),
+            Self::Ttou => wrap(libc::SIGTTOU),
+            Self::Urg => wrap(libc::SIGURG),
+            Self::Usr1 => wrap(libc::SIGUSR1),
+            Self::Usr2 => wrap(libc::SIGUSR2),
+            Self::Vtalrm => wrap(libc::SIGVTALRM),
+            Self::Winch => wrap(libc::SIGWINCH),
+            Self::Xcpu => wrap(libc::SIGXCPU),
+            Self::Xfsz => wrap(libc::SIGXFSZ),
 
             Self::Rtmin(n) => {
                 let range = rt_range();
@@ -270,26 +270,26 @@ impl Name {
         #[allow(unreachable_patterns)]
         match number {
             // Standard signals
-            nix::libc::SIGABRT => Some(Self::Abrt),
-            nix::libc::SIGALRM => Some(Self::Alrm),
-            nix::libc::SIGBUS => Some(Self::Bus),
-            nix::libc::SIGCHLD => Some(Self::Chld),
-            nix::libc::SIGCONT => Some(Self::Cont),
-            nix::libc::SIGFPE => Some(Self::Fpe),
-            nix::libc::SIGHUP => Some(Self::Hup),
-            nix::libc::SIGILL => Some(Self::Ill),
-            nix::libc::SIGINT => Some(Self::Int),
-            nix::libc::SIGKILL => Some(Self::Kill),
-            nix::libc::SIGPIPE => Some(Self::Pipe),
-            nix::libc::SIGQUIT => Some(Self::Quit),
-            nix::libc::SIGSEGV => Some(Self::Segv),
-            nix::libc::SIGSTOP => Some(Self::Stop),
-            nix::libc::SIGTERM => Some(Self::Term),
-            nix::libc::SIGTSTP => Some(Self::Tstp),
-            nix::libc::SIGTTIN => Some(Self::Ttin),
-            nix::libc::SIGTTOU => Some(Self::Ttou),
-            nix::libc::SIGUSR1 => Some(Self::Usr1),
-            nix::libc::SIGUSR2 => Some(Self::Usr2),
+            libc::SIGABRT => Some(Self::Abrt),
+            libc::SIGALRM => Some(Self::Alrm),
+            libc::SIGBUS => Some(Self::Bus),
+            libc::SIGCHLD => Some(Self::Chld),
+            libc::SIGCONT => Some(Self::Cont),
+            libc::SIGFPE => Some(Self::Fpe),
+            libc::SIGHUP => Some(Self::Hup),
+            libc::SIGILL => Some(Self::Ill),
+            libc::SIGINT => Some(Self::Int),
+            libc::SIGKILL => Some(Self::Kill),
+            libc::SIGPIPE => Some(Self::Pipe),
+            libc::SIGQUIT => Some(Self::Quit),
+            libc::SIGSEGV => Some(Self::Segv),
+            libc::SIGSTOP => Some(Self::Stop),
+            libc::SIGTERM => Some(Self::Term),
+            libc::SIGTSTP => Some(Self::Tstp),
+            libc::SIGTTIN => Some(Self::Ttin),
+            libc::SIGTTOU => Some(Self::Ttou),
+            libc::SIGUSR1 => Some(Self::Usr1),
+            libc::SIGUSR2 => Some(Self::Usr2),
 
             // Non-standard but common signals
             #[cfg(any(
@@ -304,15 +304,15 @@ impl Name {
                 target_os = "nto",
                 target_os = "solaris",
             ))]
-            nix::libc::SIGPOLL => Some(Self::Poll),
-            nix::libc::SIGPROF => Some(Self::Prof),
-            nix::libc::SIGSYS => Some(Self::Sys),
-            nix::libc::SIGTRAP => Some(Self::Trap),
-            nix::libc::SIGURG => Some(Self::Urg),
-            nix::libc::SIGVTALRM => Some(Self::Vtalrm),
-            nix::libc::SIGWINCH => Some(Self::Winch),
-            nix::libc::SIGXCPU => Some(Self::Xcpu),
-            nix::libc::SIGXFSZ => Some(Self::Xfsz),
+            libc::SIGPOLL => Some(Self::Poll),
+            libc::SIGPROF => Some(Self::Prof),
+            libc::SIGSYS => Some(Self::Sys),
+            libc::SIGTRAP => Some(Self::Trap),
+            libc::SIGURG => Some(Self::Urg),
+            libc::SIGVTALRM => Some(Self::Vtalrm),
+            libc::SIGWINCH => Some(Self::Winch),
+            libc::SIGXCPU => Some(Self::Xcpu),
+            libc::SIGXFSZ => Some(Self::Xfsz),
 
             // other signals
             #[cfg(not(any(
@@ -323,7 +323,7 @@ impl Name {
                 target_os = "linux",
                 target_os = "redox",
             )))]
-            nix::libc::SIGEMT => Some(Self::Emt),
+            libc::SIGEMT => Some(Self::Emt),
             #[cfg(not(any(
                 target_os = "aix",
                 target_os = "android",
@@ -333,7 +333,7 @@ impl Name {
                 target_os = "linux",
                 target_os = "redox",
             )))]
-            nix::libc::SIGINFO => Some(Self::Info),
+            libc::SIGINFO => Some(Self::Info),
             #[cfg(any(
                 target_os = "aix",
                 target_os = "android",
@@ -345,9 +345,9 @@ impl Name {
                 target_os = "nto",
                 target_os = "solaris",
             ))]
-            nix::libc::SIGIO => Some(Self::Io),
+            libc::SIGIO => Some(Self::Io),
             #[cfg(target_os = "horizon")]
-            nix::libc::SIGLOST => Some(Self::Lost),
+            libc::SIGLOST => Some(Self::Lost),
             #[cfg(any(
                 target_os = "aix",
                 target_os = "android",
@@ -359,7 +359,7 @@ impl Name {
                 target_os = "redox",
                 target_os = "solaris",
             ))]
-            nix::libc::SIGPWR => Some(Self::Pwr),
+            libc::SIGPWR => Some(Self::Pwr),
             #[cfg(all(
                 any(
                     target_os = "android",
@@ -369,9 +369,9 @@ impl Name {
                 ),
                 not(any(target_arch = "mips", target_arch = "mips64", target_arch = "sparc64"))
             ))]
-            nix::libc::SIGSTKFLT => Some(Self::Stkflt),
+            libc::SIGSTKFLT => Some(Self::Stkflt),
             #[cfg(target_os = "freebsd")]
-            nix::libc::SIGTHR => Some(Self::Thr),
+            libc::SIGTHR => Some(Self::Thr),
 
             // Real-time signals
             _ => {
@@ -414,9 +414,9 @@ fn all_signals() -> impl Iterator<Item = Number> {
 /// Converts the signal set to a vector of signal numbers.
 ///
 /// This function adds the signal numbers in the set to the vector.
-pub(super) fn sigset_to_vec(set: *const nix::libc::sigset_t, vec: &mut Vec<Number>) {
+pub(super) fn sigset_to_vec(set: *const libc::sigset_t, vec: &mut Vec<Number>) {
     vec.extend(
-        all_signals().filter(|number| unsafe { nix::libc::sigismember(set, number.as_raw()) == 1 }),
+        all_signals().filter(|number| unsafe { libc::sigismember(set, number.as_raw()) == 1 }),
     );
 }
 
@@ -426,18 +426,18 @@ impl Disposition {
     /// This function returns the `sigaction` in an `MaybeUninit` because the
     /// `sigaction` structure may contain platform-dependent extra fields that
     /// are not initialized by this function.
-    pub(super) fn to_sigaction(self) -> MaybeUninit<nix::libc::sigaction> {
+    pub(super) fn to_sigaction(self) -> MaybeUninit<libc::sigaction> {
         let handler = match self {
-            Disposition::Default => nix::libc::SIG_DFL,
-            Disposition::Ignore => nix::libc::SIG_IGN,
+            Disposition::Default => libc::SIG_DFL,
+            Disposition::Ignore => libc::SIG_IGN,
             Disposition::Catch => super::catch_signal as *const extern "C" fn(c_int) as _,
         };
 
-        let mut sa = MaybeUninit::<nix::libc::sigaction>::uninit();
+        let mut sa = MaybeUninit::<libc::sigaction>::uninit();
         let sa_ptr = sa.as_mut_ptr();
         unsafe {
-            (&raw mut (*sa_ptr).sa_flags).write(0);
-            nix::libc::sigemptyset(&raw mut ((*sa_ptr).sa_mask));
+            (*sa_ptr).sa_flags = 0;
+            libc::sigemptyset(&raw mut (*sa_ptr).sa_mask);
 
             #[cfg(not(target_os = "aix"))]
             #[allow(clippy::useless_transmute)] // See from_sigaction below
@@ -451,18 +451,18 @@ impl Disposition {
     }
 
     /// Converts the `sigaction` to the signal disposition for the real system.
-    pub(super) unsafe fn from_sigaction(sa: &MaybeUninit<nix::libc::sigaction>) -> Self {
+    pub(super) unsafe fn from_sigaction(sa: &MaybeUninit<libc::sigaction>) -> Self {
         #[cfg(not(target_os = "aix"))]
-        let handler = (&raw const (*sa.as_ptr()).sa_sigaction).read();
+        let handler = (*sa.as_ptr()).sa_sigaction;
 
         #[cfg(target_os = "aix")]
-        let handler = (&raw const (*sa.as_ptr()).sa_union.__su_sigaction).read();
+        let handler = (*sa.as_ptr()).sa_union.__su_sigaction;
 
         // It is platform-specific whether we really need to transmute the handler.
         #[allow(clippy::useless_transmute)]
         match std::mem::transmute(handler) {
-            nix::libc::SIG_DFL => Self::Default,
-            nix::libc::SIG_IGN => Self::Ignore,
+            libc::SIG_DFL => Self::Default,
+            libc::SIG_IGN => Self::Ignore,
             _ => Self::Catch,
         }
     }