+++ yash-env/CHANGELOG.md
@@ -13,12 +13,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The `builtin::Builtin` struct now has the `is_declaration_utility` field.
 - The `builtin::Builtin` struct now can be constructed with the associated
   function `new`.
+- The `system::errno::Errno` struct now can be converted to and from the `Errno`
+  type from the `errno` crate.
+- Internal dependencies:
+    - errno 0.3.10
+    - libc 0.2.169
 
 ### Changed
 
+- `System::getpwnam_dir` now takes a `&CStr` parameter instead of a `&str`.
 - The `builtin::Builtin` struct is now `non_exhaustive`.
 - External dependency versions:
     - yash-syntax 0.13.0 → 0.14.0
+
+### Removed
+
+- The implementation of `From` for converting `errno::Errno` to and from
+  `nix::errno::Errno`.
+- Internal dependencies:
+    - nix 0.29.0
 
 ## [0.5.0] - 2024-12-14